Update project.tex
line 143 - change in the Trout population is 0 since we're graphing the rate of change of the Trout population as a function of the Trout population.
Edited to: rate of change int he trout population is 0.

line 145 - Below direction fields for each Trout species is modelled.
Edited to: Direction fields for each Trout species are modelled below

line 167 - The Brown trout the only species with 4 equilibrium solutions, 2 stable, 2 unstable
Edited to: The Brown Trout is the only species with 4 equilibrium solutions, 2 stable, 2 unstable:

potential changes,
line 174 - label the equilibrium solution for rainbow trout. Also if zero is mentioned as a equil. solns. for all species, should we not specify it for the brown trout?